backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
pymongo==4.6.0
motor==3.3.2

# HTTP and Web Scraping
aiohttp==3.9.1
async-timeout==4.0.3
beautifulsoup4==4.12.2
certifi==2023.11.17

# Machine Learning and AI
transformers==4.36.2
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Data Validation and Models
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and Monitoring
structlog==23.2.0